feat: Adiciona MonthNavigator
Altera DiarioNavigator para controlar a navegação pe página por mês
